refactor(l1): move P2P blocks into a separate module.
